added Asset loading support to iOS

+++ src/mix/mix_string.cpp
@@ -10,7 +10,7 @@ StringFormatter::StringFormatter (uint32_t _initialBufferSize)
 }
 
 Utf8Buffer::Utf8Buffer (const char* _utf8Str)
-    : Buffer (strlen (_utf8Str) + 1, (const uint8_t*)_utf8Str)
+    : Buffer ((uint32_t)strlen (_utf8Str) + 1, (const uint8_t*)_utf8Str)
 {
 }
